Move ssh helpers into new file
